backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-PyJWT==1.0.0
Flask-PyMongo==3.0.1
Flask-JWT-Extended==4.5.3
bcrypt==4.2.1
pymongo==4.11.1
gridfs==2.0.2
whisper==1.1.10
pydub==0.25.1
openai==1.2.3
python-dotenv==1.0.1
werkzeug==3.1.3
